move OptionModel class to its own file
