+++ src/Tkmm/ViewModels/Pages/TotKOptimizerPageViewModel.cs
@@ -1,14 +1,10 @@
-using System;
-using System.Collections.Generic;
 using System.Collections.ObjectModel;
-using System.IO;
-using System.Linq;
 using System.Text;
 using System.Text.Json;
-using System.Threading.Tasks;
 using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
 using Tkmm.Core;
+using Tkmm.Models;
 
 namespace Tkmm.ViewModels.Pages;
 
@@ -52,7 +48,7 @@ public partial class TotKOptimizerPageViewModel : ObservableObject
 
         foreach (var option in Options)
         {
-            if (option.ConfigClass == null || option.ConfigClass.Count < 2)
+            if (option.ConfigClass.Count < 2)
                 continue;
 
             string section = option.ConfigClass[0];
@@ -225,49 +221,4 @@ public partial class TotKOptimizerPageViewModel : ObservableObject
 
         return options;
     }
-}
-
-public class OptionModel : ObservableObject
-{
-    public string Name { get; set; }
-    public string DefaultValue { get; set; }
-    public List<string> Values { get; set; }
-    public List<string> NameValues { get; set; }
-
-    private string selectedValue;
-    public string SelectedValue
-    {
-        get => selectedValue;
-        set => SetProperty(ref selectedValue, value);
-    }
-
-    private int selectedIndex;
-    public int SelectedIndex
-    {
-        get => selectedIndex;
-        set
-        {
-            if (SetProperty(ref selectedIndex, value))
-            {
-                if (Class != null && Class.ToLowerInvariant() == "dropdown" && value >= 0 && value < Values.Count)
-                {
-                    SelectedValue = Values[value];
-                    OnPropertyChanged(nameof(DisplaySelectedValue));
-                }
-            }
-        }
-    }
-
-    public string DisplaySelectedValue =>
-        (Class?.ToLowerInvariant() == "dropdown" && NameValues != null && NameValues.Count > SelectedIndex)
-            ? NameValues[SelectedIndex]
-            : SelectedValue;
-
-    public string Class { get; set; }
-    public string Section { get; set; }
-    public double Increments { get; set; }
-    public List<string> ConfigClass { get; set; }
-    public bool Auto { get; set; }
-
-    public IEnumerable<string> DisplayValues => (NameValues?.Count ?? 0) > 0 ? NameValues : Values;
 }
